update layout swap form price alert
